revise preprocess

+++ api.py
@@ -22,8 +22,8 @@ class APIData(BaseModel):
     loan_intent :str
     loan_grade :str
     loan_amnt : int
-    loan_int_rate : int  
-    loan_percent_income : int
+    loan_int_rate : float  
+    loan_percent_income : float
     cb_person_default_on_file :str
     cb_person_cred_hist_length : int
